11-02

+++ 04-02-2020/EX-revisao/src/br/com/fiap/bean/ContaInvestimento.java
@@ -2,7 +2,6 @@ package br.com.fiap.bean;
 
 public interface ContaInvestimento {
 
-	public float calculaRetornoInvestimento();
+	public double calculaRetornoInvestimento();
 
-	
 }
